requirements.txt
streamlit
Pillow
pandas
jinja2
reportlab
streamlit-drawable-canvas
